Moved scripts, tests use multiprocessing not subprocess
